🚚 Rename: provider 파일명. 컨벤션에 맞춰 변경
