+++ src/components/providers/toast-provider.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
 import { useMemo, useState } from 'react';
 
-import type { ToastData } from '@/types/toast';
+import type { ToastData, ToastMessage } from '@/types/toast';
 
 import Toast from '@/components/progress-toast';
 
@@ -10,7 +10,7 @@ import { ToastContext } from '@/context/toast';
 export function ToastProvider({ children }: { children: ReactNode }) {
   const [toast, setToast] = useState<ToastData | null>(null);
 
-  const addToast = (message: string, type: 'success' | 'error' | 'info' | 'warning', duration: number = 5000) => {
+  const addToast = (message: string, type: ToastMessage, duration: number = 5000) => {
     const newToast = { id: Date.now(), message, type, duration };
     setToast(newToast);
   };
