+++ src/pages/_app.tsx
@@ -3,8 +3,8 @@ import type { NextPage } from 'next';
 import type { AppProps } from 'next/app';
 import { SessionProvider } from 'next-auth/react';
 
-import ReactQueryProvider from '@/components/providers/ReactQuery';
-import { ToastProvider } from '@/components/providers/ToastContext';
+import ReactQueryProvider from '@/components/providers/react-query-provider';
+import { ToastProvider } from '@/components/providers/toast-provider';
 
 import '@/styles/globals.scss';
